+++ frontend/pages/ManageControlsPage/MacOSUpdates/MacOSUpdates.tsx
@@ -1,4 +1,5 @@
 import React, { useContext } from "react";
+import { InjectedRouter } from "react-router";
 
 import { AppContext } from "context/app";
 
@@ -9,15 +10,17 @@ import PremiumFeatureMessage from "components/PremiumFeatureMessage";
 
 import OsMinVersionForm from "./components/OsMinVersionForm";
 import NudgePreview from "./components/NudgePreview";
+import TurnOnMdmMessage from "../components/TurnOnMdmMessage/TurnOnMdmMessage";
 
 const baseClass = "mac-os-updates";
 
 interface IMacOSUpdates {
+  router: InjectedRouter;
   teamIdForApi: number;
 }
 
-const MacOSUpdates = ({ teamIdForApi }: IMacOSUpdates) => {
-  const { isPremiumTier } = useContext(AppContext);
+const MacOSUpdates = ({ router, teamIdForApi }: IMacOSUpdates) => {
+  const { config, isPremiumTier } = useContext(AppContext);
 
   const OperatingSystemCard = useInfoCard({
     title: "macOS versions",
@@ -34,6 +37,10 @@ const MacOSUpdates = ({ teamIdForApi }: IMacOSUpdates) => {
       />
     ),
   });
+
+  if (!config?.mdm.enabled_and_configured) {
+    return <TurnOnMdmMessage router={router} />;
+  }
 
   return isPremiumTier ? (
     <div className={baseClass}>
