+++ frontend/pages/ManageControlsPage/components/TurnOnMdmMessage/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./TurnOnMdmMessage";
